third_party/blink/web_tests/platform/generic/http/tests/devtools/profiler/heap-snapshot-containment-shown-node-count-preserved-when-sorting-expected.txt

Tests Containment view of detailed heap snapshots. Shown node count must be preserved after sorting.

Profiler was enabled.
Detailed heap profiles were enabled.

Running: testShownNodeCountPreservedWhenSorting

Profiler was disabled.
